Refactor run_inference_engine script
Changed the way the run_inference_engine script handles a given input
dataset. Also refactored the evaluation logic given the changes to the
structure of the ground truth dataset.

Also added a script for splitting and preparing HDFS datasets for train,
validation and test sets when a ground truth dataset is available.