Change the way the feature transformation is created
Now uses a dataset containing ALL possible log keys for a given system
instead of just the log keys in the training data. This accounts for all
possible classes. Also moves toward supporting the idea of a datastore.

Output size of the Anomaly Detection LSTM model must now be specified in
the inference engine configuration file.