Refactor the way the feature extractor gets the number of unique keys
The number, and identifiers, of the possible log keys are no longer just
generated from the training dataset. Instead, these are extracted from
the log event template matrix that the Data Miner outputs. This gives a
more complete representation of the log key space. The template matrix
must be specified in the configuration file for the inference engine.